requirements.txt
google-adk==1.28.0
google-genai>=1.64.0,<2.0.0
google-auth>=2.47.0,<3.0.0
fastapi>=0.124.1,<1.0.0
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
pg8000==1.31.2
pydantic>=2.12.0,<3.0.0
python-dotenv==1.1.0